+++ CHANGES.rst
@@ -41,6 +41,7 @@ require additional help please ask on `/r/redditdev
 * :meth:`praw.models.Front.random_rising`,
   :meth:`praw.models.Subreddit.random_rising` and
   :meth:`praw.models.Multireddit.random_rising`.
+* :class:`~.WikiPage` supports a revision argument.
 * Support installed-type OAuth apps.
 * Support read-only OAuth for all application types.
 * Support script-type OAuth apps.